// src/types/stats.ts
export interface Stats {
    total_users: number;
    total_companies: number;
    total_forms: number;
    total_questions: number;
    total_answers: number;
    total_templates_all: number;
    total_templates_workers: number;
    total_templates_owned_facility: number;
    total_templates_products: number;
    total_templates_facility: number;
    total_onboarding: number;
    total_accepted_onboarding: number;
    total_rejected_onboarding: number;
  }